Move Span/ReadOnlySpan to shared CoreLib partition (#10988)
Fix a few method names to better names used in CoreRT

Contributes to https://github.com/dotnet/corert/issues/2966
